chore: split samples into individual files (#622)
* chore: split samples into individual files

* chore: removed cli option

* chore: update copyright

* fix: update to latest format

* fix: tests

* fix: broken tests
