+++ dialogflow/resource.js
@@ -18,103 +18,6 @@
 // Operations for intents
 // /////////////////////////////////////////////////////////////////////////////
 
-async function listIntents(projectId) {
-  // [START dialogflow_list_intents]
-  // Imports the Dialogflow library
-  const dialogflow = require('@google-cloud/dialogflow');
-
-  // Instantiates clients
-  const intentsClient = new dialogflow.IntentsClient();
-
-  // The path to identify the agent that owns the intents.
-  const projectAgentPath = intentsClient.agentPath(projectId);
-
-  const request = {
-    parent: projectAgentPath,
-  };
-
-  console.log(projectAgentPath);
-
-  // Send the request for listing intents.
-  const [response] = await intentsClient.listIntents(request);
-  response.forEach(intent => {
-    console.log('====================');
-    console.log(`Intent name: ${intent.name}`);
-    console.log(`Intent display name: ${intent.displayName}`);
-    console.log(`Action: ${intent.action}`);
-    console.log(`Root folowup intent: ${intent.rootFollowupIntentName}`);
-    console.log(`Parent followup intent: ${intent.parentFollowupIntentName}`);
-
-    console.log('Input contexts:');
-    intent.inputContextNames.forEach(inputContextName => {
-      console.log(`\tName: ${inputContextName}`);
-    });
-
-    console.log('Output contexts:');
-    intent.outputContexts.forEach(outputContext => {
-      console.log(`\tName: ${outputContext.name}`);
-    });
-  });
-  // [END dialogflow_list_intents]
-}
-
-async function createIntent(
-  projectId,
-  displayName,
-  trainingPhrasesParts,
-  messageTexts
-) {
-  // [START dialogflow_create_intent]
-  // Imports the Dialogflow library
-  const dialogflow = require('@google-cloud/dialogflow');
-
-  // Instantiates the Intent Client
-  const intentsClient = new dialogflow.IntentsClient();
-
-  // The path to identify the agent that owns the created intent.
-  const agentPath = intentsClient.agentPath(projectId);
-
-  const trainingPhrases = [];
-
-  trainingPhrasesParts.forEach(trainingPhrasesPart => {
-    const part = {
-      text: trainingPhrasesPart,
-    };
-
-    // Here we create a new training phrase for each provided part.
-    const trainingPhrase = {
-      type: 'EXAMPLE',
-      parts: [part],
-    };
-
-    trainingPhrases.push(trainingPhrase);
-  });
-
-  const messageText = {
-    text: messageTexts,
-  };
-
-  const message = {
-    text: messageText,
-  };
-
-  const intent = {
-    displayName: displayName,
-    trainingPhrases: trainingPhrases,
-    messages: [message],
-  };
-
-  const createIntentRequest = {
-    parent: agentPath,
-    intent: intent,
-  };
-
-  // Create the intent
-  const responses = await intentsClient.createIntent(createIntentRequest);
-  console.log(`Intent ${responses[0].name} created`);
-  // [END dialogflow_create_intent]
-}
-
 async function deleteIntent(projectId, intentId) {
   // [START dialogflow_delete_intent]
   // Imports the Dialogflow library
@@ -264,45 +167,6 @@ const cli = require('yargs')
   .boolean('force')
   .alias('force', ['f'])
   .describe('force', 'force operation without a prompt')
-  .command(
-    'create-intent',
-    'Create Intent',
-    {
-      displayName: {
-        alias: 'd',
-        string: true,
-        demandOption: true,
-        requiresArg: true,
-        description: 'Display Name',
-      },
-      trainingPhrasesParts: {
-        alias: 't',
-        array: true,
-        string: true,
-        demandOption: true,
-        requiresArg: true,
-        description: 'Training Phrases',
-      },
-      messageTexts: {
-        alias: 'm',
-        array: true,
-        string: true,
-        demandOption: true,
-        requiresArg: true,
-        description: 'Message Texts',
-      },
-    },
-    opts =>
-      createIntent(
-        opts.projectId,
-        opts.displayName,
-        opts.trainingPhrasesParts,
-        opts.messageTexts
-      )
-  )
-  .command('list-intents', 'List Intent', {}, opts =>
-    listIntents(opts.projectId)
-  )
   .command(
     'delete-intent',
     'Delete Intent',
