+++ dialogflow/system-test/create-intent.test.js
@@ -18,24 +18,20 @@ const {assert} = require('chai');
 const {after, describe, it} = require('mocha');
 const execSync = require('child_process').execSync;
 const uuid = require('uuid');
+const projectId =
+  process.env.GCLOUD_PROJECT || process.env.GOOGLE_CLOUD_PROJECT;
 const dialogflow = require('@google-cloud/dialogflow');
 
 const exec = cmd => execSync(cmd, {encoding: 'utf8'});
 
 describe('create intent', () => {
   const client = new dialogflow.IntentsClient();
-  const cmd = 'node resource.js';
+  const cmd = 'node create-intent.js';
   const displayName = `fake_display_name_${uuid.v4().split('-')[0]}`;
-  const phrase1 = 'training_phrase_1';
-  const phrase2 = 'training_phrase_2';
-  const message1 = 'message_1';
-  const message2 = 'message_2';
   let intentId;
 
   it('should create an intent', async () => {
-    const output = exec(
-      `${cmd} create-intent -d ${displayName} -t ${phrase1} -t ${phrase2} -m ${message1} -m ${message2}`
-    );
+    const output = exec(`${cmd} ${projectId} ${displayName}`);
     assert.include(output, 'intents');
     intentId = output.split(' ')[1].split('/')[4];
   });
